feat and refactor: finish and check timesheet endpoint route and delete unused file

+++ src/reports/progress/progress.service.ts
@@ -1,6 +1,10 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ConflictException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { ProgressRepository } from './progress.repository';
-import { ProgressReport } from '@prisma/client';
+import { Prisma, ProgressReport } from '@prisma/client';
 import { ReportQueryDto } from 'src/reports/dto/query-report.dto';
 import { CustomResponse } from 'src/_common/res/response';
 import { CreateProgressDto } from './dto/create-progress.dto';
@@ -27,19 +31,39 @@ export class ProgressService {
   }
 
   async create(data: CreateProgressDto): Promise<ProgressReport> {
-    return await this.progressRepository.create({
-      ...data,
-      user: { connect: { id: data.userId } },
-      project: { connect: { id: data.projectId } },
-      phase: { connect: { id: data.pjtPhaseId } },
-    });
+    try {
+      return await this.progressRepository.create({
+        ...data,
+        user: { connect: { id: data.userId } },
+        project: { connect: { id: data.projectId } },
+        phase: { connect: { id: data.pjtPhaseId } },
+      });
+    } catch (error) {
+      if (error instanceof Prisma.PrismaClientKnownRequestError) {
+        if (error.code === 'P2002') {
+          // P2002 = unique constraint violation
+          throw new ConflictException(
+            `Timesheet for user ${data.userId} on project ${data.projectId} at ${data.reportDate} already exists`
+          );
+        }
+      }
+      throw error;
+    }
   }
 
   async update(id: number, data: UpdateProgressDto): Promise<ProgressReport> {
+    const isExist = await this.progressRepository.findOne(id);
+    if (!isExist) {
+      throw new NotFoundException(`Progress Report #${id} is not found`);
+    }
     return await this.progressRepository.update(id, data);
   }
 
   async remove(id: number): Promise<CustomResponse> {
+    const isExist = await this.progressRepository.findOne(id);
+    if (!isExist) {
+      throw new NotFoundException(`Progress Report #${id} is not found`);
+    }
     return await this.progressRepository.remove(id);
   }
 }
